### 2025-07-04 ba90230
Textual and grammar improvements
