+++ paper/paper.bib
@@ -30,7 +30,8 @@ url="https://doi.org/10.1007/978-1-4614-7657-3_19"
   author={Kirillov, Alexander and Mintun, Eric and Ravi, Nikhila and Mao, Hanzi and Rolland, Chloe and Gustafson, Laura and Xiao, Tete and Whitehead, Spencer and Berg, Alexander C and Lo, Wan-Yen and Dollar, Piotr and Girshick, Ross},
   booktitle={Proceedings of the IEEE/CVF international conference on computer vision},
   pages={4015--4026},
-  year={2023}
+  year={2023},
+  doi={10.1109/ICCV51070.2023.00371}
 }
 
 @article{ravi2024sam,
@@ -47,7 +48,8 @@ url="https://doi.org/10.1007/978-1-4614-7657-3_19"
   journal={Advances in Neural Information Processing Systems},
   volume={37},
   pages={110746--110783},
-  year={2024}
+  year={2024},
+  doi={10.48550/arXiv.2311.13385}
 }
 
 @article{wang2023sam,
@@ -97,7 +99,8 @@ month = dec,
 title = {{MITK}},
 url = {https://github.com/MITK/MITK},
 version = {v2024.12},
-year = {2024}
+year = {2024},
+doi = {10.5281/zenodo.14356025}
 }
 
 @article{isensee2021nnu,
@@ -127,5 +130,6 @@ year = {2024}
   title={Kidney abnormality segmentation in thorax-abdomen CT scans},
   author={Mamani, Gabriel Efrain Humpire and Lessmann, Nikolas and Scholten, Ernst Th and Prokop, Mathias and Jacobs, Colin and van Ginneken, Bram},
   journal={arXiv preprint arXiv:2309.03383},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2309.03383}
 }
