### 2025-09-14 #8276 @Kevin-Mattheus-Moerman
@liu-yikang, @pchlap, @ibro45, @surajpaib, @vasudev-sharma, @NMontanaBrown, @suyashkumar, @zapaishchykova, @cnmy-ro, @ObedDzik, @thewtex  would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled _SlicerNNInteractive: A 3D Slicer extension for nnInteractive_.

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/coendevente/SlicerNNInteractive) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08276/joss.08276/10.21105.joss.08276.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of a completed review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!
