### 2025-09-16 #9003 @editorialbot
[REVIEW]: SlicerNNInteractive: A 3D Slicer extension for nnInteractive
**Submitting author:** @coendevente (<a href="http://orcid.org/0000-0001-5908-8367">Coen de Vente</a>)
**Repository:** https://github.com/coendevente/SlicerNNInteractive
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.1
