### 2025-09-18 ae203d7
Dockerfile update for compatibility with newest Miniconda version
