### 2025-09-18 4fdbeb2
added uv instructions to rereadme
