### 2025-11-01 d7db0fe
Small textual update in README.md
