### 2025-11-05 49714c5
Made tests robust to when screen DPI is not 1.0
