### 2025-11-05 be22f21
Fixed non ascii characters
