+++ timing_results/plot_timing.py
@@ -6,11 +6,11 @@ Make a stacked-bar comparison (prompt-server vs local processing)
 for high- vs low-compute measurements stored in two JSON files.
 
 The script expects:
-	1) high-compute JSON path
-	2) low-compute  JSON path
+    1) high-compute JSON path
+    2) low-compute  JSON path
 optional:
-	--out <file>      image file to write (PNG, PDF, etc.)
-	--dpi <int>       resolution (default 150)
+    --out <file>    image file to write (PNG, PDF, etc.)
+    --dpi <int>     resolution (default 150)
 """
 
 import json, argparse, os, sys
@@ -26,201 +26,201 @@ mpl.rcParams.update({"font.family": "Arial", "font.size": 12.9})
 MODES = ["point", "bbox", "scribble", "lasso"]
 
 LABELS = {
-	"point":    "Point",
-	"bbox":     "Bounding box",   # ← the wording you want
-	"scribble": "Scribble",
-	"lasso":    "Lasso",
+    "point":    "Point",
+    "bbox":     "Bounding box",
+    "scribble": "Scribble",
+    "lasso":    "Lasso",
 }
 SIZES = ["S", "M", "L"]
-COLOR_CYCLE = plt.rcParams["axes.prop_cycle"].by_key()["color"][:4]  # C0–C3
+COLOR_CYCLE = plt.rcParams["axes.prop_cycle"].by_key()["color"][:4]  # C0-C3
 
 
 def load_stats(path: Path):
-	"""
-	Return nested dict stats[size][mode] with means / stds (seconds).
-	"""
-	with open(path, "r", encoding="utf-8") as f:
-		raw = json.load(f)
+    """
+    Return nested dict stats[size][mode] with means / stds (seconds).
+    """
+    with open(path, "r", encoding="utf-8") as f:
+        raw = json.load(f)
 
-	stats = {}
-	for size in SIZES:
-		stats[size] = {}
-		for mode in MODES:
-			prompt = np.array(
-				[x["server_prompt_processing"] for x in raw[size][mode]], dtype=float
-			) / 1000.0  # -> seconds
-			local = np.array(
-				[x["remainder"] for x in raw[size][mode]], dtype=float
-			) / 1000.0
-			total = prompt + local
-			stats[size][mode] = {
-				"prompt_mean": prompt.mean(),
-				"prompt_std": prompt.std(ddof=1),
-				"local_mean": local.mean(),
-				"local_std": local.std(ddof=1),
-				"total_mean": total.mean(),
-				"total_std": total.std(ddof=1),
-			}
-	return stats
+    stats = {}
+    for size in SIZES:
+        stats[size] = {}
+        for mode in MODES:
+            prompt = np.array(
+                [x["server_prompt_processing"] for x in raw[size][mode]], dtype=float
+            ) / 1000.0  # -> seconds
+            local = np.array(
+                [x["remainder"] for x in raw[size][mode]], dtype=float
+            ) / 1000.0
+            total = prompt + local
+            stats[size][mode] = {
+                "prompt_mean": prompt.mean(),
+                "prompt_std": prompt.std(ddof=1),
+                "local_mean": local.mean(),
+                "local_std": local.std(ddof=1),
+                "total_mean": total.mean(),
+                "total_std": total.std(ddof=1),
+            }
+    return stats
 
 
 def build_figure(stats_high, stats_low, out_file: Path, dpi: int):
-	fig, ax = plt.subplots(figsize=(10, 6), dpi=dpi)
+    fig, ax = plt.subplots(figsize=(10, 6), dpi=dpi)
 
-	bar_width = 0.18
-	x = np.arange(len(SIZES) * 2)  # S-L, M-L, L-L, S-H, M-H, L-H
[…]
