### 2025-11-05 #9003 @thewtex
@coendevente great 👍 . It would be helpful for reproducibility and quality to have a test on the `main` branch that triggers a segmentation on test data and verifies the output.
