### 2025-11-11 be3e546
Feature: Add macOS MPS support
