### 2025-11-17 a639e7d
Added regression tests
