### 2025-11-18 513c07b
Minor textual update in readme
