### 2025-12-10 d00fca7
Added test files to CMakeLists.txt
