### 2025-12-11 #9003 @coendevente
Yes, I looked into the error. Apologies for the late response; it took some time to figure out what went wrong. There was an issue related to the CMake files, which didn't occur while developing and testing through the Developer Extension (only when installing from the Extensions Manager). I expect the new build will be in the Extensions Manager by tomorrow.
