### 2025-12-24 e2988e2
Updated print statements
