### 2026-01-23 #9003 @coendevente
Just a quick update: I've looked into the two issues. The first one has been closed, and the fix for the second one is ready to be tested by @thewtex.
