### 2026-02-10 #9003 @coendevente
@Kevin-Mattheus-Moerman Yes, that's okay with me. The current status is that both issues are addressed and completed from my side. I was just waiting for @thewtex to confirm the second issue is solved on their side as well.

Happy to assist with any remaining issues if needed, @thewtex!
